+++ apps/app/package.json
@@ -10,7 +10,6 @@
     "test": "vitest run"
   },
   "dependencies": {
-    "@arcjet/next": "1.0.0-alpha.28",
     "@clerk/nextjs": "^6.3.0",
     "@repo/analytics": "workspace:*",
     "@repo/database": "workspace:*",
@@ -18,6 +17,7 @@
     "@repo/env": "workspace:*",
     "@repo/feature-flags": "workspace:*",
     "@repo/next-config": "workspace:*",
+    "@repo/security": "workspace:*",
     "@repo/tailwind-config": "workspace:*",
     "@sentry/nextjs": "^8.37.1",
     "lucide-react": "^0.456.0",
