+++ docs/security.mdx
@@ -26,7 +26,7 @@ import { Authors } from '/snippets/authors.mdx';
 [Arcjet bot detection](https://docs.arcjet.com/bot-protection/concepts) is configured to prevent scrapers and other automated clients. You can [configure additional bot types](https://docs.arcjet.com/bot-protection/identifying-bots) to allow or block. [Arcjet Shield WAF](https://docs.arcjet.com/shield/concepts) will detect and block common attacks such as SQL injection, cross-site scripting, and other OWASP Top 10 vulnerabilities.
 
 - For the website, bot detection and the Arcjet Shield WAF are both configured in `apps/web/middleware.ts` to block scrapers and other bots, but still allow search engines and preview link generators. This will run on every request by default, except for static assets.
-- For the app, there is a central Arcjet client configured at `apps/app/lib/arcjet.ts` which includes the Arcjet Shield WAF rules. This client is extended in the authenticated route layout in `apps/app/app/(authenticated)/layout.tsx` with bot detection to block all bots except preview link generators. This will run just on authenticated routes. Clerk includes bot detection and rate limiting in their login route handlers by default.
+- For the app, there is a central Arcjet client configured at `@repo/security` which includes the Arcjet Shield WAF rules. This client is extended in the authenticated route layout in `apps/app/app/(authenticated)/layout.tsx` with bot detection to block all bots except preview link generators. This will run just on authenticated routes. Clerk includes bot detection and rate limiting in their login route handlers by default.
 
 When a rule is triggered, the request will be blocked and an error returned. You can customize the error message in code, redirect to a different page, or handle the error in a different way as needed.
